Fixing Issue #85 (Duplicated Move Data)
Contains a fix and testing for the following issue:

I found that we are sometimes sending multiple exact duplicates of
pokemon move data when there are different levels at which a pokemon
learns a move for different version groups.  I found that the problem
was that we were using order_by('level') and duplicate('move_id') .
This link explains why this behavior happens.
https://docs.djangoproject.com/en/1.7/ref/models/querysets/#distinct