fix: Remove old website and associated code
The pokeapi.co website has been rebuilt and moved to a separate
project, so the website in this project is obsolete.

This removes all files and code related to the website, including the
'alerts' and 'hits' apps (which managed website alerts and the hit
counter) and unused site assets (CSS, JS, images).