Changed Scale Shot Effect Text
The effect text for Scale Shot (move ID 799) was modified to be more accurate and specific, in accordance with Issue #749.

As part of doing so, 3 files were modified:
1. move_effects.csv
The new effect_id 443 was added.

2. moves.csv
The effect_id of scale_shot was changed from 1 to 443.

3. move_effect_prose.csv
The new effect text under the ID 443 was added.